+++ jpos/src/main/java/org/jpos/space/SpaceUtil.java
@@ -78,7 +78,7 @@ public class SpaceUtil {
             Object obj = sp.inp (key);
             wipe (sp, key); // just in case
             if (obj instanceof Long) 
-                l = ((Long)obj).longValue();
+                l = (Long) obj;
             sp.out (key, ++l);
         }
         return l;
